requirements.txt
absl-py==0.13.0
aio-pika==6.8.0
aiofiles==0.8.0
aiogram==2.25.1
aiohttp==3.7.4
aiormq==3.3.1
aiosignal==1.3.1
APScheduler==3.7.0
asgiref==3.4.1
asn1crypto==1.5.1
aspose-words==23.5.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
asynctest==0.13.0
attrdict==2.0.1
attrs==21.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bidict==0.21.4
blinker==1.6.2
blis==0.7.5
boto3==1.20.24
botocore==1.23.24
CacheControl==0.12.10
cached-property==1.5.2
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==3.1.0
clang==5.0
click==8.0.3
cloudpickle==1.6.0
coincurve==15.0.1
colorama==0.4.4
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
comtypes==1.1.14
confluent-kafka==1.9.2
cryptography==3.4.8
cycler==0.11.0
cymem==2.0.6
cytoolz==0.12.1
dask==2021.11.2
decorator==5.1.0
Django==3.2.10
django-cors-headers==3.10.1
dm-tree==0.1.6
dnspython==1.16.0
docopt==0.6.2
docx==0.2.4
docx2txt==0.8
eciespy==0.3.10
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fbmessenger==6.0.0
fire==0.4.0
Flask==2.0.1
Flask-Mail==0.9.1
flatbuffers==1.12
frozenlist==1.3.3
fsspec==2021.11.1
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.42.0
h5py==3.1.0
hexbytes==0.3.0
httptools==0.3.0
humanfriendly==10.0
idna==2.10
importlib-metadata==4.8.2
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.0.1
json2html==1.3.0
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langcodes==3.3.0
libclang==16.0.0
locket==0.2.1
lru-dict==1.1.8
lxml==4.9.2
magic-filter==1.0.9
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mattermostwrapper==2.2
msgpack==1.0.3
multidict==5.2.0
murmurhash==1.0.6
mysql-connector==2.2.9
mysqlclient==2.0.1
networkx==2.6.3
numpy==1.21.6
oauthlib==3.1.1
opencv-contrib-python==4.3.0.36
opencv-python==4.1.1.26
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pandas==1.1.5
parsimonious==0.8.1
partd==1.2.0
pathy==0.6.1
pbkdf2==1.3
pickle4==0.0.1
Pillow==8.4.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
portalocker==2.7.0
preshed==3.0.6
prompt-toolkit==2.0.10
protobuf==3.19.1
psycopg2-binary==2.9.2
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.18.0
pydantic==1.8.2
pydot==1.4.2
PyJWT==2.1.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.6
pypiwin32==223
pypng==0.0.21
PyQRCode==1.2.1
pyreadline==2.1
pyrsistent==0.18.0
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-engineio==4.3.0
python-socketio==5.5.0
pyttsx3==2.90
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0
questionary==1.10.0
randomname==0.1.5
rasa==3.0.2
rasa-sdk==3.0.2
redis==3.5.3
regex==2021.8.28
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rlp==2.0.1
rocketchat-API==1.16.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
sanic==21.9.3
Sanic-Cors==1.0.1
sanic-jwt==1.7.0
sanic-plugin-toolkit==1.2.0
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.3.1
six==1.15.0
sklearn-crfsuite==0.3.6
slack-sdk==3.21.3
slackclient==2.9.3
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.4.28
sqlparse==0.4.2
srsly==2.4.2
tabulate==0.8.9
tarsafe==0.0.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.1
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.13.0
termcolor==1.1.0
terminaltables==3.1.10
thinc==8.0.13
threadpoolctl==3.0.0
toolz==0.11.2
tqdm==4.62.3
twilio==6.50.1
typeguard==2.13.3
typer==0.4.0
typing-extensions==3.7.4.3
typing-utils==0.1.0
tzdata==2023.3
tzlocal==2.1
ujson==4.3.0
urllib3==1.26.7
wasabi==0.9.0
wcwidth==0.2.5
web3==4.7.2
webexteamssdk==1.6
websockets==10.1
Werkzeug==2.0.2
wrapt==1.12.1
yarl==1.7.2
zipp==3.6.0
